refactor(cli): extract native-hooks group into its own module
Move the `native-hooks` Click group and its `register` subcommand out
of the 5.9k-line cli.py monolith into src/dopemux/commands/native_hooks_commands.py,
matching the established pattern (instances_commands.py, etc.).
cli.py now imports and registers the group via add_command().

Behavior preserved: `dopemux native-hooks register [--global]` still
writes the same hook entry to the same settings.json, resolving the
native hook script through Path(__file__).parent.parent (adjusted for
the new module depth).
