+++ src/dopemux/cli.py
@@ -2516,86 +2516,8 @@ def _trigger_dope_context_autoindex_startup(
 from .commands.instances_commands import instances
 cli.add_command(instances, "instances")
 
-@cli.group("native-hooks")
-def native_hooks():
-    """
-    🔗 Protocol Synchronization: Manage Claude Code internal hooks
-
-    Orchestrates the registration and management of high-fidelity internal 
-    hooks. These rituals ensure that Claude Code activity is seamlessly 
-    synchronized with the DØPEMÜX cockpit telemetry.
-    """
-    pass
-
-@native_hooks.command("register")
-@click.option("--global", "is_global", is_flag=True, help="🌐 Global Calibration: Register ritual hooks in the global configuration ledger.")
-def native_hooks_register(is_global: bool):
-    """
-    ⚡ Synchronize Protocol: Register DØPEMÜX hooks in Claude settings
-
-    Automates the injection of ritual hook coordinates into the Claude 
-    Code configuration ledger, enabling real-time signal detection.
-    """
-    import json
-    from pathlib import Path
-    
-    # Path to this script's native hook entry point
-    hook_script = Path(__file__).resolve().parent / "claude" / "native_hooks.py"
-    cmd = f"python3 {hook_script}"
-    
-    # Define hook configuration
-    hooks_config = {
-        "hooks": {
-            "command": [
-                {
-                    "events": [
-                        "SessionStart", 
-                        "UserPromptSubmit", 
-                        "PreToolUse", 
-                        "PermissionRequest", 
-                        "PostToolUse", 
-                        "PostToolUseFailure", 
-                        "Stop", 
-                        "SubagentStop", 
-                        "PreCompact", 
-                        "SessionEnd"
-                    ],
-                    "command": cmd
-                }
-            ]
-        }
-    }
-    
-    # Target settings file
-    if is_global:
-        settings_path = Path.home() / ".claude" / "settings.json"
-    else:
-        settings_path = Path.cwd() / ".claude" / "settings.json"
-        
-    # Read existing settings
-    existing = {}
-    if settings_path.exists():
-        try:
-            existing = json.loads(settings_path.read_text())
-        except:
-            pass
-            
-    # Simple merge (Dopemux hooks first)
-    if "hooks" not in existing:
-        existing["hooks"] = {}
-    if "command" not in existing["hooks"]:
-        existing["hooks"]["command"] = []
-        
-    # Check if already registered
-    already_registered = any(h.get("command") == cmd for h in existing["hooks"]["command"])
-    
-    if not already_registered:
-        existing["hooks"]["command"].insert(0, hooks_config["hooks"]["command"][0])
-        settings_path.parent.mkdir(parents=True, exist_ok=True)
-        settings_path.write_text(json.dumps(existing, indent=2))
-        console.print(f"[success]✓ Registered Dopemux native hooks in {settings_path}[/success]")
-    else:
-        console.print(f"[info]Dopemux native hooks already registered in {settings_path}[/info]")
+from .commands.native_hooks_commands import native_hooks
+cli.add_command(native_hooks, "native-hooks")
 
 
 @cli.command(name="pr-merge", context_settings=dict(ignore_unknown_options=True))
